Add websites - Batch 1
Welcome to the stage where the websites are added.
This is Batch 1.

Side note, if anyone knows how to make a json file for the links and have the javascript read it and randomly choose one then redirect to that site, then please make a pull request. That would mean a lot to me!